requirements.txt
wykop-sdk-reborn>=0.3.0
requests~=2.24.0
urllib3~=1.25.11
pillow~=8.0.1
pytest==6.1.2